+++ lib/my_banner.rb
@@ -14,6 +14,7 @@ require "my_banner/calendar_service"
 require "my_banner/schedule"
 require "my_banner/schedule/tableset"
 require "my_banner/section"
+require "my_banner/section/meeting"
 
 module MyBanner
 end
